+++ src/app/search/search.controller.js
@@ -42,11 +42,11 @@
                     'Active': true,
                     'Retired': false,
                     'Suspended by ONC-ACB': true,
-                    'Withdrawn by Developer': true,
-                    'Withdrawn by Developer Under Surveillance/Review': true,
-                    'Withdrawn by ONC-ACB': true,
+                    'Withdrawn by Developer': false,
+                    'Withdrawn by Developer Under Surveillance/Review': false,
+                    'Withdrawn by ONC-ACB': false,
                     'Suspended by ONC': true,
-                    'Terminated by ONC': true
+                    'Terminated by ONC': false
                 },
                 certificationEdition: {
                     '2011': false,